updated security check, fixed file upload validation

+++ src/AppBundle/Controller/RequirementController.php
@@ -4,9 +4,11 @@ namespace AppBundle\Controller;
 
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 use Symfony\Component\HttpFoundation\Request;
 use AppBundle\Entity\Requirement;
 use Symfony\Component\HttpFoundation\JsonResponse;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\Method;
 
 /**
  * @Route("reikalavimai")
@@ -16,7 +18,6 @@ class RequirementController extends Controller
     /**
      * @param Requirement $requirement
      *
-     * @return JsonResponse
      * @Route("/{id}",
      *     options = { "expose" = true },
      *     name = "delete_requirement",
@@ -24,10 +25,13 @@ class RequirementController extends Controller
      */
     public function deleteAction(Requirement $requirement)
     {
+        $user = $this->getUser();
+        if (! $user || ! $user->hasRole('ROLE_USER_EMPLOYER')) {
+            throw new AccessDeniedException();
+        }
+
         $em = $this->getDoctrine()->getManager();
         $em->remove($requirement);
         $em->flush();
-
-        return new JsonResponse();
     }
 }
